Module spinoza::utils

An assortment of utility functions for visualizing, benchmarking, and testing.

Functions

assert_float_closeness

Asserts that two floating point numbers are approximately equal.

gen_random_state

Generates a random quantum state

mat_mul_2x2

Utility function for multiplying two 2 x 2 gates

padded_bin

Convert a usize to its binary expansion, but padded with 0’s. Padding is of size, width.

pretty_print_int

Formats an unsigned, 128 bit integer with commas, as a string. Used for readability

swap

Swap using controlled X gates.

to_table

Display the State as a table