Crate spinoza

The rust core of Spinoza

Modules

circuit

Abstractions for a quantum circuit

config

Configuration options for running spinoza

consts

Consts for Quantum State Simulation, such as the Pauli Gates and the Hadamard gate

core

Abstractions for representing a Quantum State

gates

Abstractions for quantum logic gates

math

An assortment of mathematical structures, functions, and constants for quantum state simulation.

measurement

Functionality for measurement

openqasm

Functionality for creating a QuantumCircuit from an OpenQASM 2.0 program

unitaries

Functionality for applying large 2^n * 2^n matrices to the state Ideally, this should be a last resort

utils

An assortment of utility functions for visualizing, benchmarking, and testing.